Improve export by handling it server side (#41)
* WIP - unified export

* Handle export server side

With these changes, export is now handled server side. Specifically,
the excel/csv file is generated on the server then send to the client.
This greatly improves performance & allows us to overcome any
maxResult limit.

On the frontend a new base ExportButton is created to provide a
consistent look & feel as well as handling the request to the server.
On the backend, an export endpoint has been created. It takes the format
(csv/excel) and any filters then produce the corresponding file before
sending it back to the client.

* Code cleanup

---------

Co-authored-by: Prince Duepa <pd74280@uga.edu>